+++ Tests/AwesomeAssertions.Equivalency.Specs/MemberMatchingSpecs.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Diagnostics.CodeAnalysis;
+using System.Linq;
 using JetBrains.Annotations;
 using Xunit;
 using Xunit.Sdk;
@@ -358,6 +359,62 @@ public class MemberMatchingSpecs
             .NotBeEquivalentTo(expectation, opt => opt
                 .ExcludingMissingMembers()
                 .WithMapping("Property2", "Property1"));
+    }
+
+    [Fact]
+    public void Exclusion_can_be_configured_only_once()
+    {
+        var subject = new[] { 1 };
+        var expectation = new[] { 2 };
+
+        Action act = () => subject.Should().BeEquivalentTo(expectation, o => o
+            .ExcludingMissingMembers().ExcludingMissingMembers());
+
+        var message = act.Should().Throw<XunitException>().Which.Message;
+        var configurationMessages =
+            message.Split("\n".ToCharArray(), StringSplitOptions.RemoveEmptyEntries).Select(x => x.Trim());
+        configurationMessages.GroupBy(x => x).Should().OnlyContain(x => x.Count() == 1);
+    }
+
+    [Fact]
+    public void Throwing_can_be_configured_only_once()
+    {
+        var subject = new[] { 1 };
+        var expectation = new[] { 2 };
+
+        Action act = () => subject.Should().BeEquivalentTo(expectation, o => o
+            .ThrowingOnMissingMembers().ThrowingOnMissingMembers());
+
+        var message = act.Should().Throw<XunitException>().Which.Message;
+        var configurationMessages =
+            message.Split("\n".ToCharArray(), StringSplitOptions.RemoveEmptyEntries).Select(x => x.Trim());
+        configurationMessages.GroupBy(x => x).Should().OnlyContain(x => x.Count() == 1);
+    }
+
+    [Fact]
+    public void Exclusion_after_throwing_let_exclude_win()
+    {
+        var subject = new[] { 1 };
+        var expectation = new[] { 2 };
+
+        Action act = () => subject.Should().BeEquivalentTo(expectation, o => o
+            .ThrowingOnMissingMembers().ExcludingMissingMembers());
+
+        act.Should().Throw<XunitException>().WithMessage("*Try to match member by name*")
+            .And.Message.Should().NotContain("Match member by name (or throw)");
+    }
+
+    [Fact]
+    public void Throwing_after_exclude_let_throwing_win()
+    {
+        var subject = new[] { 1 };
+        var expectation = new[] { 2 };
+
+        Action act = () => subject.Should().BeEquivalentTo(expectation, o => o
+            .ExcludingMissingMembers().ThrowingOnMissingMembers());
+
+        act.Should().Throw<XunitException>().WithMessage("*Match member by name (or throw)*")
+            .And.Message.Should().NotContain("Try to match member by name");
     }
 
     [Fact]